SomeOptionsAdded

+++ StudyManagement.Domain/OrderAgg/IOrderRepository.cs
@@ -5,10 +5,8 @@ namespace StudyManagement.Domain.OrderAgg
 {
     public interface IOrderRepository : IRepositoryBase<long , Order>
     {
-        double GetAmountBy(long id);
+        int GetAmountBy(long id);
         List<OrderViewModel> Search(OrderSearchModel searchModel);
         List<OrderItemViewModel> GetItems(long orderId);
-
-
     }
 }
